requirements.txt
pandas
numpy
seaborn
matplotlib
mysql-connector-python
streamlit
python-dotenv
sqlalchemy